+++ plugins/ucc_chat/lib/ucc_chat/models/schema/pinned_message.ex
@@ -2,6 +2,9 @@ defmodule UccChat.Schema.PinnedMessage do
   use UccChat.Shared, :schema
 
   alias UccChat.Schema.{Message, Channel}
+  alias UcxUcc.UccPubSub
+
+  require Logger
 
   schema "pinned_messages" do
     belongs_to :message, Message
@@ -22,5 +25,24 @@ defmodule UccChat.Schema.PinnedMessage do
     |> cast(params, @fields)
     |> validate_required(@fields)
     |> unique_constraint(:message_id, name: :pinned_messages_channel_id_message_id)
+    |> prepare_changes(&prepare_notify/1)
+  end
+
+  defp prepare_notify(%{action: :insert} = changeset) do
+    channel_id = changeset.changes[:channel_id]
+    UccPubSub.broadcast "pin:insert", "channel:#{channel_id}" ,
+      %{channel_id: channel_id}
+    changeset
+  end
+
+  defp prepare_notify(%{action: :delete} = changeset) do
+    channel_id = changeset.data.channel_id
+    UccPubSub.broadcast "pin:delete", "channel:#{channel_id}" ,
+      %{channel_id: channel_id}
+    changeset
+  end
+
+  defp prepare_notify(changeset) do
+    changeset
   end
 end