+++ plugins/ucc_chat/lib/ucc_chat/models/schema/starred_message.ex
@@ -3,6 +3,7 @@ defmodule UccChat.Schema.StarredMessage do
 
   alias UcxUcc.Accounts.User
   alias UccChat.Schema.{Message, Channel}
+  alias UcxUcc.UccPubSub
 
   schema "starred_messages" do
     belongs_to :user, User
@@ -24,5 +25,24 @@ defmodule UccChat.Schema.StarredMessage do
     |> cast(params, @fields)
     |> validate_required(@fields)
     |> unique_constraint(:user_id, name: :starred_messages_user_id_channel_id_message_id)
+    |> prepare_changes(&prepare_notify/1)
+  end
+
+  defp prepare_notify(%{action: :insert} = changeset) do
+    channel_id = changeset.changes[:channel_id]
+    UccPubSub.broadcast "star:insert", "channel:#{channel_id}" ,
+      %{channel_id: channel_id}
+    changeset
+  end
+
+  defp prepare_notify(%{action: :delete} = changeset) do
+    channel_id = changeset.data.channel_id
+    UccPubSub.broadcast "star:delete", "channel:#{channel_id}" ,
+      %{channel_id: channel_id}
+    changeset
+  end
+
+  defp prepare_notify(changeset) do
+    changeset
   end
 end